==> 2. Clean Amazon Data.ipynb <==
import numpy as np
import pandas as pd
import re
import time


df = pd.read_csv('1. product_laptop.csv')
df.head(1)

### Data Preprocessing

df.shape

df.duplicated().sum()

df[df.duplicated(keep=False)]

df.drop_duplicates(inplace=True)

df.shape

df.duplicated(subset=["Individual Rating", "Review Title", "Review Date", "Product Description", 'Review Text'], keep=False).sum()

df[df.duplicated(subset=["Individual Rating", "Review Title", "Review Date", "Product Description", 'Review Text'], keep=False)] \
.sort_values(by=["Review Title", "Review Date", "Product Description", 'Review Text']).head(10)

df.drop_duplicates(subset=["Individual Rating", "Review Title", "Review Date", "Product Description", 'Review Text'], keep='first', inplace=True)

df.shape

df.isnull().sum()

df.info()

df['Price'].nunique()

df['Price'].unique()

df['Price'] = df['Price'].str.replace(",", "")
df['Price'] = df['Price'].str.strip('.')
df['Price'] = df['Price'].astype(dtype = int)
df['Price'].head(2)

df['Rating Count'].nunique()

df['Rating Count'].unique()

df['Rating Count'] = df['Rating Count'].str.replace(",", "")
df['Rating Count'] = df['Rating Count'].astype(dtype = 'Int64')
df['Rating Count'].head(2)

df['Review Count'].nunique()

df['Review Count'].unique()

df['Review Count'] = df['Review Count'].str.replace(",", "")
df['Review Count'] = df['Review Count'].astype(dtype = 'Int64')
df['Review Count'].head(2)

df['Bought L Month'].unique()

df['Helpful Votes'].nunique()

df['Helpful Votes'].unique()

def func(x):
    if isinstance(x, str) and x.lower()=='one':
        return '1'
    else:
        return x

df['Helpful Votes'] = df['Helpful Votes'].apply(func)

df['Helpful Votes'].unique()

df['Helpful Votes'] = df['Helpful Votes'].astype(dtype="Int64")
df['Helpful Votes']

df.info()

df['ASIN'].nunique()

df['ASIN'].unique()

## ASIN column is clean

df['Generic Name'].nunique()

df['Generic Name'].unique()

df[(df['Generic Name'].str.contains('laptop', case = False, na=False)) | (df['Generic Name'].str.contains('macbook', case = False, na=False))]

df[~(df['Generic Name'].str.contains('laptop|macbook|chromebook', case = False, na=False))]

df[~(df['Generic Name'].str.contains('laptop|macbook|chromebook', case = False, na=False))]['Generic Name'].nunique()

df[~(df['Generic Name'].str.contains('laptop|macbook|chromebook', case = False, na=False))]['Generic Name'].unique()

df[df['Generic Name']=="Mobile Workstation"]['Product Title'].values

df[(~df['Generic Name'].str.contains('laptop|macbook|chromebook', case = False, na=False)) & \
(~df['Product Title'].str.contains('laptop|macbook|chromebook', case = False, na=False))]['Generic Name'].nunique()

df[(~df['Generic Name'].str.contains('laptop|macbook|chromebook', case = False, na=False)) & \
(~df['Product Title'].str.contains('laptop|macbook|chromebook', case = False, na=False))]['Generic Name'].unique()

df.shape

exclude_idx = df[
    (~df['Generic Name'].str.contains('laptop|macbook|chromebook', case=False, na=False)) &
    (~df['Product Title'].str.contains('laptop|macbook|chromebook', case=False, na=False))
].index

df = df[~df.index.isin(exclude_idx)].reset_index(drop=True)


df.shape

df.isnull().sum()

def clean_text(text):
    if pd.isna(text):
        return None

    text = re.sub(r'[^\x00-\x7F]+', '', text)

    text = re.sub(r'[^\w\s@#.,!?:/\-]', '', text)

    text = re.sub(r'\s+', ' ', text).strip()

    if len(text)<10:
        return None

    return text.lower()

df['Cleaned Review Text'] = df['Review Text'].apply(clean_text)

df['Cleaned Review Text'].head(2)

df.shape

df.isnull().sum()

df.dropna(subset=['Cleaned Review Text'], inplace=True)

df.shape

df.isnull().sum()

df = df.reset_index(drop=True)
df.insert(0, "Index", range(1, len(df) + 1))

# df.to_csv('2. Cleaned_laptop_df.csv', index=False)
